Fix role for newest release of ansible.
Manual checks are now disabled via default value, as the feature we used
before was deprecated and has now been removed. "ansible_devices" is now
a dict of dicts, so we need to slightly change our handling of it for
the SMART check. This requires a few lines of python in filter_plugins
and therefore changes to the playbooks.